Pull main to payroll branch

+++ src/main/java/com/worknest/features/leave/dto/LeaveBalanceDto.java
@@ -1,11 +1,12 @@
 package com.worknest.features.leave.dto;
 
 import com.worknest.domain.enums.LeaveType;
+import java.math.BigDecimal;
 
 public record LeaveBalanceDto(
         LeaveType leaveType,
-        int totalDays,
-        int usedDays,
-        int availableDays
+        BigDecimal totalDays,
+        BigDecimal usedDays,
+        BigDecimal availableDays
 ) {
 }
